+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## dev ##
+
+### enhancements
+  * SimpleForm support for multiple attributes. (by [@nashby](https://github.com/nashby))
+  * Array-like multiple attributes. (by [@lest](https://github.com/lest))
+
 ## 0.4.0 (September 6, 2012) ##
 
 Legacy support was dropped. The following versions are supported: